1: continue complete the Surf Logic 2: continue cover the unit tests for Surf Logic

+++ src/main/java/com/cybercafe/model/WeekPrice.java
@@ -16,6 +16,6 @@ public class WeekPrice {
     private Long memberType;
     private Long ignoreTime;
     private Float startPrice;
-    private Float minCostPrice;
+    private Integer minCostPrice;
     private String price;
 }
